move utility function to utils
